use new pos calc to ensure correct above and below pos
rancher/rancher#13535
