+++ lib/shared/addon/components/action-menu/component.js
@@ -2,6 +2,7 @@ import Component from '@ember/component';
 import layout from './template';
 import { inject as service } from '@ember/service'
 import { computed, get } from '@ember/object';
+import calculatePosition from 'shared/utils/calculate-position';
 
 
 export default Component.extend({
@@ -39,6 +40,8 @@ export default Component.extend({
       get(this, 'tooltipService').set('childOpened', false);
       get(this, 'tooltipService').hide();
     },
+
+    calculatePosition: calculatePosition,
 
   },
