+++ lib/shared/addon/components/action-menu/template.hbs
@@ -1,6 +1,6 @@
 {{#basic-dropdown
      horizontalPosition="right"
-     verticalPosition="below"
+     calculatePosition=(action "calculatePosition" )
      onOpen=(action 'actionsOpen')
      onClose=(action 'actionsClosed')
      as |dd|
