+++ lib/shared/app/calculate-position/util.js
@@ -1,0 +1,1 @@
+export { default } from 'shared/calculate-position/util';
